bidderUtils/getBidIdParameter
