+++ libraries/bidderUtils/bidderUtils.js
@@ -1,0 +1,7 @@
+export function getBidIdParameter(key, paramsObj) {
+  if (paramsObj && paramsObj[key]) {
+    return paramsObj[key];
+  }
+
+  return '';
+}
