+++ modules/adgenerationBidAdapter.js
@@ -1,8 +1,9 @@
-import {tryAppendQueryString, getBidIdParameter, escapeUnsafeChars, deepAccess} from '../src/utils.js';
+import {tryAppendQueryString, escapeUnsafeChars, deepAccess} from '../src/utils.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {BANNER, NATIVE} from '../src/mediaTypes.js';
 import {config} from '../src/config.js';
 import {convertOrtbRequestToProprietaryNative} from '../src/native.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const ADG_BIDDER_CODE = 'adgeneration';
