+++ modules/admaticBidAdapter.js
@@ -1,7 +1,8 @@
-import { getValue, logError, isEmpty, deepAccess, getBidIdParameter, isArray } from '../src/utils.js';
+import { getValue, logError, isEmpty, deepAccess, isArray } from '../src/utils.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { config } from '../src/config.js';
 import { BANNER, VIDEO } from '../src/mediaTypes.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 let SYNC_URL = '';
 const BIDDER_CODE = 'admatic';
 export const spec = {
