+++ modules/audiencerunBidAdapter.js
@@ -2,7 +2,6 @@ import {
   _each,
   deepAccess,
   formatQS,
-  getBidIdParameter,
   getValue,
   isArray,
   isFn,
@@ -12,6 +11,7 @@ import {
 import {config} from '../src/config.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {BANNER} from '../src/mediaTypes.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'audiencerun';
 const BASE_URL = 'https://d.audiencerun.com';
