+++ modules/beopBidAdapter.js
@@ -1,7 +1,6 @@
 import {
   buildUrl,
   deepAccess,
-  getBidIdParameter,
   getValue,
   isArray,
   logInfo,
@@ -12,6 +11,7 @@ import {getRefererInfo} from '../src/refererDetection.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {config} from '../src/config.js';
 import {getAllOrtbKeywords} from '../libraries/keywords/keywords.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'beop';
 const ENDPOINT_URL = 'https://hb.beop.io/bid';
