+++ modules/bidglassBidAdapter.js
@@ -1,6 +1,7 @@
-import { _each, isArray, getBidIdParameter, deepClone, getUniqueIdentifierStr } from '../src/utils.js';
+import { _each, isArray, deepClone, getUniqueIdentifierStr } from '../src/utils.js';
 // import {config} from 'src/config.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'bidglass';
