+++ modules/cadentApertureMXBidAdapter.js
@@ -1,7 +1,6 @@
 import {
   _each,
   deepAccess,
-  getBidIdParameter,
   isArray,
   isFn,
   isPlainObject,
@@ -14,6 +13,7 @@ import {BANNER, VIDEO} from '../src/mediaTypes.js';
 import {Renderer} from '../src/Renderer.js';
 import {find, includes} from '../src/polyfill.js';
 import {parseDomain} from '../src/refererDetection.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'cadent_aperture_mx';
 const ENDPOINT = 'hb.emxdgt.com';
