+++ modules/ebdrBidAdapter.js
@@ -1,6 +1,7 @@
-import { logInfo, getBidIdParameter } from '../src/utils.js';
+import { logInfo} from '../src/utils.js';
 import { VIDEO, BANNER } from '../src/mediaTypes.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 const BIDDER_CODE = 'ebdr';
 export const spec = {
   code: BIDDER_CODE,
