+++ modules/getintentBidAdapter.js
@@ -1,5 +1,6 @@
-import { getBidIdParameter, isFn, isInteger } from '../src/utils.js';
+import { isFn, isInteger } from '../src/utils.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'getintent';
 const IS_NET_REVENUE = true;
