+++ modules/gmosspBidAdapter.js
@@ -2,7 +2,6 @@ import {
   createTrackPixelHtml,
   deepAccess,
   deepSetValue,
-  getBidIdParameter,
   getDNT,
   getWindowTop,
   isEmpty,
@@ -12,6 +11,7 @@ import {
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {config} from '../src/config.js';
 import {BANNER} from '../src/mediaTypes.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'gmossp';
 const ENDPOINT = 'https://sp.gmossp-sp.jp/hb/prebid/query.ad';
