+++ modules/growadvertisingBidAdapter.js
@@ -1,9 +1,10 @@
 'use strict';
 
-import { getBidIdParameter, deepAccess, _each, triggerPixel } from '../src/utils.js';
+import { deepAccess, _each, triggerPixel } from '../src/utils.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {BANNER, NATIVE} from '../src/mediaTypes.js';
 import { convertOrtbRequestToProprietaryNative } from '../src/native.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'growads';
