+++ modules/mgidBidAdapter.js
@@ -9,7 +9,6 @@ import {
   isEmpty,
   triggerPixel,
   logWarn,
-  getBidIdParameter,
   isFn,
   isNumber,
   isBoolean,
@@ -21,6 +20,7 @@ import {config} from '../src/config.js';
 import { getStorageManager } from '../src/storageManager.js';
 import { convertOrtbRequestToProprietaryNative } from '../src/native.js';
 import {USERSYNC_DEFAULT_CONFIG} from '../src/userSync.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const GVLID = 358;
 const DEFAULT_CUR = 'USD';
