+++ modules/nextMillenniumBidAdapter.js
@@ -2,7 +2,6 @@ import {
   _each,
   createTrackPixelHtml,
   deepAccess,
-  getBidIdParameter,
   getDefinedParams,
   getWindowTop,
   isArray,
@@ -20,6 +19,7 @@ import * as events from '../src/events.js';
 
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {getRefererInfo} from '../src/refererDetection.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'nextMillennium';
 const ENDPOINT = 'https://pbs.nextmillmedia.com/openrtb2/auction';
