+++ modules/open8BidAdapter.js
@@ -1,8 +1,9 @@
 import { Renderer } from '../src/Renderer.js';
 import {ajax} from '../src/ajax.js';
-import { createTrackPixelHtml, getBidIdParameter, logError, logWarn, tryAppendQueryString } from '../src/utils.js';
+import { createTrackPixelHtml, logError, logWarn, tryAppendQueryString } from '../src/utils.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { VIDEO, BANNER } from '../src/mediaTypes.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'open8';
 const URL = 'https://as.vt.open8.com/v1/control/prebid';
