+++ modules/orbitsoftBidAdapter.js
@@ -1,5 +1,6 @@
 import * as utils from '../src/utils.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BIDDER_CODE = 'orbitsoft';
 let styleParamsMap = {
@@ -45,10 +46,10 @@ export const spec = {
     for (let i = 0; i < validBidRequests.length; i++) {
       bidRequest = validBidRequests[i];
       let bidRequestParams = bidRequest.params;
-      let placementId = utils.getBidIdParameter('placementId', bidRequestParams);
-      let requestUrl = utils.getBidIdParameter('requestUrl', bidRequestParams);
-      let referrer = utils.getBidIdParameter('ref', bidRequestParams);
-      let location = utils.getBidIdParameter('loc', bidRequestParams);
+      let placementId = getBidIdParameter('placementId', bidRequestParams);
+      let requestUrl = getBidIdParameter('requestUrl', bidRequestParams);
+      let referrer = getBidIdParameter('ref', bidRequestParams);
+      let location = getBidIdParameter('loc', bidRequestParams);
       // Append location & referrer
       if (location === '') {
         location = utils.getWindowLocation();
@@ -58,7 +59,7 @@ export const spec = {
       }
 
       // Styles params
-      let stylesParams = utils.getBidIdParameter('style', bidRequestParams);
+      let stylesParams = getBidIdParameter('style', bidRequestParams);
       let stylesParamsArray = {};
       for (let currentValue in stylesParams) {
         if (stylesParams.hasOwnProperty(currentValue)) {
@@ -74,7 +75,7 @@ export const spec = {
         }
       }
       // Custom params
-      let customParams = utils.getBidIdParameter('customParams', bidRequestParams);
+      let customParams = getBidIdParameter('customParams', bidRequestParams);
       let customParamsArray = {};
       for (let customField in customParams) {
         if (customParams.hasOwnProperty(customField)) {
