+++ modules/trionBidAdapter.js
@@ -1,6 +1,7 @@
-import { getBidIdParameter, parseSizesInput, tryAppendQueryString } from '../src/utils.js';
+import { parseSizesInput, tryAppendQueryString } from '../src/utils.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import { getStorageManager } from '../src/storageManager.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const BID_REQUEST_BASE_URL = 'https://in-appadvertising.com/api/bidRequest';
 const USER_SYNC_URL = 'https://in-appadvertising.com/api/userSync.html';
