+++ modules/waardexBidAdapter.js
@@ -1,8 +1,9 @@
-import {deepAccess, getBidIdParameter, isArray, logError} from '../src/utils.js';
+import {deepAccess, isArray, logError} from '../src/utils.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {BANNER, VIDEO} from '../src/mediaTypes.js';
 import {config} from '../src/config.js';
 import {find} from '../src/polyfill.js';
+import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 
 const ENDPOINT = `https://hb.justbidit.xyz:8843/prebid`;
 const BIDDER_CODE = 'waardex';
