+++ test/spec/utils_spec.js
@@ -3,6 +3,7 @@ import {expect} from 'chai';
 import CONSTANTS from 'src/constants.json';
 import * as utils from 'src/utils.js';
 import {deepEqual, memoize, waitForElementToLoad} from 'src/utils.js';
+import {getBidIdParameter} from '../../libraries/bidderUtils/bidderUtils.js';
 
 var assert = require('assert');
 
@@ -19,25 +20,6 @@ describe('Utils', function () {
     type_array = 'Array',
     type_function = 'Function';
 
-  describe('getBidIdParameter', function () {
-    it('should return value of the key in input object', function () {
-      var obj = {
-        a: 'valueA',
-        b: 'valueB'
-      };
-      var output = utils.getBidIdParameter('a', obj);
-      assert.equal(output, 'valueA');
-    });
-
-    it('should return empty string, if the key is not existsed in the object', function () {
-      var obj = {
-        a: 'valueA',
-        b: 'valueB'
-      };
-      var output = utils.getBidIdParameter('c', obj);
-      assert.equal(output, '');
-    });
-  });
 
   describe('tryAppendQueryString', function () {
     it('should append query string to existing url', function () {
